==> 00. Html Scraping.ipynb <==
# HTML Scraping of Pixar and Disney Classic Movies

#Import libaries 
import requests
import pandas as pd
from bs4 import BeautifulSoup
import re
import numpy as np 
import os

## Access the websites

# Movies from movies.fandom.com 
urls = {'Pinocchio': 'https://movies.fandom.com/wiki/Pinocchio_(1940)/Transcript',
        'One Hundred and One Dalmatians': 'https://movies.fandom.com/wiki/One_Hundred_and_One_Dalmatians/Transcript',
        'The Fox and the Hound':'https://movies.fandom.com/wiki/The_Fox_and_the_Hound/Transcript',
        'The Great Mouse Detective':'https://movies.fandom.com/wiki/The_Great_Mouse_Detective/Transcript',
        'Oliver & Company':'https://movies.fandom.com/wiki/Oliver_%26_Company/Transcript',
        'The Emperors New Groove' : 'https://movies.fandom.com/wiki/The_Emperor%27s_New_Groove/Transcript',
        'Lilo & Stitch' :'https://movies.fandom.com/wiki/Lilo_%26_Stitch/Transcript',
        'Brother Bear' : 'https://movies.fandom.com/wiki/Brother_Bear/Transcript',
        'Bolt' : 'https://movies.fandom.com/wiki/Bolt_(2008)/Transcript',
        'Frozen 2': 'https://disneyprincess.fandom.com/wiki/Frozen_2/Transcript',
        'Zootopia': 'https://zootopia.fandom.com/wiki/Zootopia/Transcript',
        'Tangled': 'https://disneyprincess.fandom.com/wiki/Tangled/Transcript',
        'Pocahontas': 'https://pocahontas.fandom.com/wiki/Pocahontas_(Disney_Movie)/Transcript',
        'The Hunchback of Notre Dame':'https://disneys-the-hunchback-of-notre-dame-wiki.fandom.com/wiki/The_Hunchback_of_Notre_Dame/Transcript',
        'The Rescuers' : 'https://movie-spoof-films.fandom.com/wiki/The_Rescuers/Transcript',
        'Fun and Fancy Free': 'https://movie-spoof-films.fandom.com/wiki/Fun_%26_Fancy_Free_(Stephen_Squirrelsky)/Transcript',
        'Cinderella':'https://disneyprincess.fandom.com/wiki/Cinderella/Transcript',
        'Moana': 'https://moana.fandom.com/wiki/Moana/Transcript',
        'Dumbo': 'https://corduroy-nelvana.fandom.com/wiki/Dumbo_(1941)_Transcript',
        'Cars': 'https://ficreation.fandom.com/wiki/Cars/Transcript',
        'Monsters, Inc.': 'https://movies.fandom.com/wiki/Monsters,_Inc./Transcript',
        'Monsters University': 'https://movies.fandom.com/wiki/Monsters_University/Transcript',
        'Toy Story 2': 'https://movies.fandom.com/wiki/Toy_Story_2/Transcript',
        'Toy Story 3': 'https://movies.fandom.com/wiki/Toy_Story_3/Transcript',
        'Toy Story 4' : 'https://movies.fandom.com/wiki/Toy_Story_4/Transcript',
        'Toy Story': 'https://movies.fandom.com/wiki/Toy_Story/Transcript',
        'Finding Nemo' : 'https://movies.fandom.com/wiki/Finding_Nemo/Transcript',
        'Finding Dory' : 'https://movies.fandom.com/wiki/Finding_Dory/Transcript',
        'Incredibles 2' :'https://the-incredibles.fandom.com/wiki/Incredibles_2/Transcript',
        'The Incredibles': 'https://the-incredibles.fandom.com/wiki/The_Incredibles/Transcript',
        'Cars 3' : 'https://geicocarinsurance.fandom.com/wiki/Cars_3/Transcript'
        
       }

print('Lenght of dict: ', len(urls))


urls_fpx = {'Sleeping Beauty': 'http://www.fpx.de/fp/Disney/Scripts/SleepingBeauty/sb.html',
        'Beauty and the Beast': 'http://www.fpx.de/fp/Disney/Scripts/BeautyAndTheBeast.txt'
       }

urls_cubbi = {'Alice in Wonderland': 'http://www.cubbi.org/disney/scripts/alice.html',
        'Hercules': 'http://www.cubbi.org/disney/scripts/hercules.txt',
        'The Lion King':'http://www.cubbi.org/disney/scripts/tlk.txt',
        'The Aristocats': 'http://www.cubbi.org/disney/scripts/aristocats.txt',
        'The Jungle Book': 'http://www.cubbi.org/disney/scripts/tjb.txt'
       }

urls_imsdb = {'Aladdin': 'https://imsdb.com/scripts/Aladdin.html',
        'Frozen': 'https://imsdb.com/scripts/Frozen-(Disney).html',
        'Mulan':'https://imsdb.com/scripts/Mulan.html',
        'The Little Mermaid':'https://imsdb.com/scripts/Little-Mermaid,-The.html',
        'The Rescuers Down Under':'https://imsdb.com/scripts/Rescuers-Down-Under,-The.html',
        'Cars 2' : 'https://imsdb.com/scripts/Cars-2.html'
       }

urls_scifiscriptscom = {'The Black Cauldron': 'http://www.scifiscripts.com/cartoon/blackcauldron.txt'
       }

# scrape the html code for movies from fandom.wiki

movie_script = pd.DataFrame()
movies = []
script = []


for key, value in urls.items():
    movie = key
    movies.append(movie)
    total_script = []
    
    url = value
    page = requests.get(url)
    soup = BeautifulSoup(page.content,'html.parser')
    results = soup.find(id="mw-content-text")
    job_elements = results.find_all("div", class_="mw-parser-output")
    total_script.append(job_elements)
    script.append(total_script)
    
    print(value)

# add movies to dataframe 
movie_script['Movie'] = movies
movie_script['Script'] = script
movie_script['Site'] = 'fandom'

# collect all scripts from FPX

movie_script_fpx = pd.DataFrame()
movies = []
script = []

for key, value in urls_fpx.items():
    movie = key
    movies.append(movie)
    total_script = []
    print(value)
    url = value
    page = requests.get(url)
    soup = BeautifulSoup(page.text, "html.parser")
    info = soup.find_all("dl", {})
    if len(info) == 0:
        info = soup.find_all("pre", {})
    total_script.append(info)
    script.append(total_script)

# add to dataframe 
movie_script_fpx['Movie'] = movies
movie_script_fpx['Script'] = script
movie_script_fpx['Site'] = 'FPX'
movie_script_fpx   


# collect all scripts from cubbi 
movie_script_cubbi = pd.DataFrame()
movies = []
script = []

for key, value in urls_cubbi.items():
    movie = key
    movies.append(movie)
    total_script = []
    print(value)
    url = value
    page = requests.get(url)
    soup = BeautifulSoup(page.text, "html.parser")
    info = soup.find_all("p", {})
    if len(info) == 0:
        info = page.text
        info = str(info)
        #remove \r','\t
        info = re.sub(r"[\r','\t]", "", info)
        index = 'THE SCRIPT'
        if index in info:
            index = info.find(index)
            info = info[index:]
        else: 
            index = 'The Script'
            if index in info:
                index = info.find('The Script')
                info = info[index:]
            else: 
                continue
        
    total_script.append(info)
    script.append(total_script)

movie_script_cubbi['Movie'] = movies
movie_script_cubbi['Script'] = script
movie_script_cubbi['Site'] = 'cubbi'


# collect all scripts from imbsd

movie_script_imsdb = pd.DataFrame()
movies = []
script = []


for key, value in urls_imsdb.items():
    movie = key
    movies.append(movie)
    total_script = []
    
    print(value)
    url = value
    page = requests.get(url)
    soup = BeautifulSoup(page.text, "html.parser")
    info = soup.find_all("pre", {})
    total_script.append(info)
    script.append(total_script)
    
movie_script_imsdb['Movie'] = movies
movie_script_imsdb['Script'] = script
movie_script_imsdb['Site'] = 'imsdb'


#collect scripts from scifiscripts.com
#The Black Cauldron  

movie_script_scifiscripts = pd.DataFrame()
movies = []
script = []

for key, value in urls_scifiscriptscom.items():
    movie = key
    movies.append(movie)
    total_script = []
    print(value)
    url = value
    page = requests.get(url)
    soup = BeautifulSoup(page.text, "html.parser")
    info = page.text
    info = str(info)
        
        
    total_script.append(info)
    script.append(total_script)

movie_script_scifiscripts['Movie'] = movies
movie_script_scifiscripts['Script'] = script
movie_script_scifiscripts['Site'] = 'scifiscripts.com'

movie_script_scifiscripts

#Peter pan get script 
movie_script_peterpan = pd.DataFrame()
movies = []
script = []



movie = 'Peter Pan'
movies.append(movie)
    
url = 'https://giveadaygetadisneyday.blogspot.com/2014/12/peter-pan-1953-script.html?fbclid=IwAR0Zlyz8BIpcpmr74GRtsI3NVLGfxODEiK2IeOIAtsgO5IrEC0hU20lgXpI'
page = requests.get(url)
soup = BeautifulSoup(page.content,'html.parser')
job_elements = soup.find_all("div", class_="region-inner main-inner")
script.append(job_elements)
    
movie_script_peterpan['Movie'] = movies
movie_script_peterpan['Script'] = script
movie_script_peterpan['Site'] = 'Blogspot'


# merge all collected scripts in one data frame 
movies_totals = [movie_script, movie_script_cubbi, movie_script_fpx,
                 movie_script_peterpan, movie_script_imsdb,
                 movie_script_scifiscripts]

movies_totals = pd.concat(movies_totals, ignore_index=True)

# convert scripts to strings 
movies_totals['Script'] = movies_totals['Script'].astype(str)

movies_totals.to_excel('Scraped_scripts.xlsx', index=False)

import pickle
def pickle_dataframe(x, filename):
    with open(filename, "wb") as f:
        pickle.dump(x, f)
        
def open_pickled_dataframe(filename):
    with open(filename, "rb") as f:
        return pickle.load(f)
    
pickle_dataframe(movies_totals, 'Scraped_scripts.pickle')

# Scraping the old snow white movie

snowwhite_url = {'Snow White and the Seven Dwarfs': 'https://the-jh-movie-collection-official.fandom.com/wiki/Snow_White_and_the_Seven_Dwarfs_(1937_film)/Transcript'
       }

# scrape the html code for movies from fandom.wiki

snowwhite_script = pd.DataFrame()
movies = []
script = []


for key, value in snowwhite_url.items():
    movie = key
    movies.append(movie)
    total_script = []
    
    url = value
    page = requests.get(url)
    soup = BeautifulSoup(page.content,'html.parser')
    results = soup.find(id="mw-content-text")
    job_elements = results.find_all("div", class_="mw-parser-output")
    total_script.append(job_elements)
    script.append(total_script)
    
    print(value)
# add movies to dataframe 
snowwhite_script['Movie'] = movies
snowwhite_script['Script'] = script
snowwhite_script['Site'] = 'fandom'
snowwhite_script['Script'][0] = str(snowwhite_script['Script'][0])

pickle_dataframe(snowwhite_script, 'snowwhite_script.pickle')
