==> 04. Creating all_scripts.ipynb <==
# Prepare analysis

#Import libaries 
import pandas as pd
from itertools import repeat
import re
import numpy as np 
import os
import matplotlib.pyplot as plt
import seaborn as sns
from statsmodels.stats.weightstats import ztest as ztest
# set directoru 
#os.getcwd()
os.chdir('/Users/Sofie/Disney Thesis')

# import clean scripts
data = pd.read_pickle("scripts_clean_final_22mar.pickle")

# we found out how there were mistakes in the disney corpus 
data.drop(data[data.Movie == "Fun and Fancy Free"].index, inplace=True)
data = data.reset_index(drop = True)

# diving the movies into 6 eras
eras = []
for index, row in data.iterrows():
    if row['Year'] < 1968:
        era = "The Golden Age"
    elif row['Year'] >= 1970 and row['Year'] <= 1988:
        era = "The Bronze Age"
    elif row['Year'] >= 1989 and row['Year'] <= 1999:
        era = "The Disney Renaissance"
    elif row['Year'] >= 2000 and row['Year'] <= 2009:
        era = "Post Renaissance Era"
    elif row['Year'] >= 2010 and row['Year'] <= 2019:
        era = "The Revival Era"
    else:
        era = "Present period"
    eras.append(era)

data['Era'] = eras

# Adding if the gender of the main character 
male = ['Bambi','Pinocchio','Fun and Fancy Free','Peter Pan','One Hundred and One Dalmatians','The Jungle Book',
       'The Aristocats','Robin Hood','The Fox and the Hound','The Black Cauldron','The Great Mouse Detective',
       'Oliver & Company','Aladdin','The Lion King','Toy Story','The Hunchback of Notre Dame','Hercules',
       "A Bug's life",'Toy Story 2','The Emperors New Groove','Atlantis','Monsters, Inc.','Finding Nemo',
       'Brother Bear','Cars','Ratatouille','Bolt','Up','Toy Story 3','Wreck-It Ralph','Monsters University',
       'The Good Dinosaur','Coco','Cars 3','Toy Story 4','Onwards','Luca','Dumbo']
female = ['Snow White and the Seven Dwarfs','Cinderella','Alice in Wonderland','Sleeping Beauty','The Little Mermaid',
         'Pocahontas','Mulan','Home on the Range','The Princess and the Frog','Tangled','Brave','Frozen',
          'Inside Out','Moana','Finding Dory','Frozen 2','Encanto','Raya and the Last Dragon']
both = ['Lady and the Tramp',
        'The Rescuers',
        'The Rescuers Down Under',
        'Lilo & Stitch',
        'The Incredibles',
        'Zootopia',
        'Incredibles 2'] 

data.loc[data['Movie'].isin(male), 'Gender_main_character'] = 'Male'
data.loc[data['Movie'].isin(female), 'Gender_main_character'] = 'Female'
data.loc[data['Movie'].isin(both), 'Gender_main_character'] = 'Both genders'


# save data 


# merge all scripts into one dataframe 
all_scripts = pd.DataFrame()

character = []
line = []
gender = []
group = []
word_count = []
movie = []
era = []

for index, row in data.iterrows():
    for _, r1 in row['Clean_Script'].iterrows():
        character.append(r1['Character_lower'])
        line.append(r1['Line'])
        gender.append(r1['Gender_updated'])
        group.append(r1['Group'])
        word_count.append(len(r1['Line'].split()))
        movie.append(row['Movie'])
        era.append(row['Era'])
        
all_scripts['Era'] = era
all_scripts['Movie'] = movie
all_scripts['Character'] = character
all_scripts['Line'] = line
all_scripts['Gender'] = gender
all_scripts['Group'] = group
all_scripts['Words_n'] = word_count

# save all_scripts 
all_scripts.to_pickle("all_scripts_cleaned_total_27apr.pickle")
data.to_pickle("scripts_clean_final_27apr.pickle")

#all_scripts_cleaned_total.pickle
